+++ include/GeoTools/MagneticTools/MagneticTile.h
@@ -5,16 +5,41 @@
 #ifndef AVIATION_CALC_UTIL_MAGNETICTILE_H
 #define AVIATION_CALC_UTIL_MAGNETICTILE_H
 
-namespace AviationCalcUtil::GeoTools::MagneticTools {
-    class MagneticTile {
+#include "GeoTools/GeoTile.h"
+#include "GeoTools/MagneticTools/MagneticResult.h"
+#include "aviationcalc_exports.h"
+#include <memory>
+#include <boost/date_time/gregorian/gregorian.hpp>
+
 #ifdef _LIBRARY
-        static vector<shared_ptr<MagneticTile>> gribTileList;
-        static mutex gribTileListLock;
-        short topLat;
-        short bottomLat;
-        short leftLon;
-        short rightLon;
+#include <vector>
+#include <mutex>
 #endif
+
+using namespace AviationCalcUtil::GeoTools;
+using namespace std;
+
+namespace AviationCalcUtil::GeoTools::MagneticTools {
+    class MagneticTile : GeoTile {
+#ifdef _LIBRARY
+    private:
+        static vector<shared_ptr<MagneticTile>> magTileList;
+        static mutex magTileListLock;
+
+        boost::gregorian::date date;
+        shared_ptr<MagneticResult> data;
+#endif
+
+    public:
+        AVIATIONCALC_EXPORT static constexpr double RESOLUTION = 0.1;
+
+        AVIATIONCALC_EXPORT static shared_ptr<const MagneticTile> findOrCreateTile(const GeoPoint &pos, const boost::gregorian::date &date);
+
+        AVIATIONCALC_EXPORT MagneticTile(const GeoPoint &point, const boost::gregorian::date &date);
+
+        AVIATIONCALC_EXPORT bool isValid(const boost::gregorian::date &date);
+
+        AVIATIONCALC_EXPORT shared_ptr<const MagneticResult> getData() const;
     };
 }